+++ docs/CHANGELOG.md
@@ -6,6 +6,12 @@ Fix:
 
   - fix NSS and GnuTLS parser: many ciphers were not parsed due to a wrong regexp
 
+Documentation:
+
+- Added a _limitations_ page
+  - No SSL support
+  - No custom cipher suites support
+
 ## [1.0.0]
 
 - First version